+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Return values of a command handler that do not have a `type` property will be ignored
+
 ### Fixed
 - Bug where event handler could not execute after Kafka retry mechanism kicked in